requirements.txt
spotipy==2.23.0
ollama==0.1.5
pyttsx3==2.90
SpeechRecognition==3.10.0
keyboard==0.13.5
pyautogui==0.9.54
pycaw==20230407
comtypes==1.2.0
screen-brightness-control==0.9.0
pyperclip==1.8.2
selenium==4.16.0
webdriver_manager==4.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.2.0
opencv-python==4.9.0.80
edge-tts==6.1.9
pygame==2.5.2
plyer==2.1.0 